Added Custom connection and example
You can now define a function for a type that can be received, in the example you can connect a color to the float "RED" and only puts the value of red in the color. If a function is not defined for a certain type, the connection cannot be created.