[Buffer] - New buffer node for all types
Templated buffer to allow storing and getting "frames" for any registered type. Created a buffer node for storing the frames, and a buffer header for getting those frames. Added default registration for default types, and added function to register custom types "registerTypeWithBufferAndHeader".
The buffer node passes it's internal buffer to the header (or multiple headers), and the header can retrieve "n" frames, and will output a vector of the type used. If T = float, then output is vector<float>.