[Connections] Improve connections system
Now it replaces the connection instead of disconnecting and reconnecting, removing setting the value to an intermediate not wanted state.
Also duplication of nodes is handled differently as you need to hold ALT during all the duplication process.
If hold ALT when disconnecting a connection, it creates a duplicate, but if you release ALT and then connect the connection, the old connection will disappear. Making a new hidden feature for better handling dynamic patching without interruptions.